requirements.txt
flask
python-dotenv
sqlalchemy
psycopg2-binary
flask_sqlalchemy
flask-cors
flask-migrate
pyjwt